# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "AttachmentBlock": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AttachmentData"
          },
          "type": {
            "const": "attachment",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "data"
        ],
        "title": "AttachmentBlock",
        "type": "object"
      },
      "AttachmentData": {
        "additionalProperties": false,
        "properties": {
          "usage_id": {
            "pattern": "^[a-z0-9][a-z0-9_-]{0,119}$",
            "title": "Usage Id",
            "type": "string"
          }
        },
        "required": [
          "usage_id"
        ],
        "title": "AttachmentData",
        "type": "object"
      },
      "CalloutBlock": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CalloutBlockData"
          },
          "type": {
            "const": "callout",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "data"
        ],
        "title": "CalloutBlock",
        "type": "object"
      },
      "CalloutBlockData": {
        "additionalProperties": false,
        "properties": {
          "markdown": {
            "minLength": 1,
            "title": "Markdown",
            "type": "string"
          },
          "tone": {
            "enum": [
              "info",
              "warning"
            ],
            "title": "Tone",
            "type": "string"
          }
        },
        "required": [
          "tone",
          "markdown"
        ],
        "title": "CalloutBlockData",
        "type": "object"
      },
...
      "CheckAndSaveRequest": {
        "additionalProperties": false,
        "properties": {
          "answer": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Answer",
            "type": "string"
          },
          "context_id": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Context Id",
            "type": "string"
          },
          "context_kind": {
            "enum": [
              "topic_lesson",
              "course_lesson",
              "standalone"
            ],
            "title": "Context Kind",
            "type": "string"
          },
          "solution_revision": {
            "minimum": 1.0,
            "title": "Solution Revision",
            "type": "integer"
          }
        },
        "required": [
          "answer",
          "solution_revision",
          "context_kind",
          "context_id"
        ],
        "title": "CheckAndSaveRequest",
        "type": "object"
      },
      "CheckAndSaveResponse": {
        "additionalProperties": false,
        "properties": {
          "correct": {
            "title": "Correct",
            "type": "boolean"
          },
          "explanation": {
            "items": {
              "discriminator": {
                "mapping": {
                  "attachment": "#/components/schemas/AttachmentBlock",
                  "callout": "#/components/schemas/CalloutBlock",
                  "code_example": "#/components/schemas/CodeExampleBlock",
                  "code_variants": "#/components/schemas/CodeVariantsBlock",
                  "completion_exercise": "#/components/schemas/WorkedExampleBlock",
                  "diagram": "#/components/schemas/DiagramBlock",
                  "image": "#/components/schemas/ImageBlock",
                  "list": "#/components/schemas/ListBlock",
                  "productive_failure_prompt": "#/components/schemas/WorkedExampleBlock",
                  "rich_text": "#/components/schemas/RichTextBlock",
                  "table": "#/components/schemas/TableBlock",
                  "text": "#/components/schemas/TextBlock",
                  "worked_example": "#/components/schemas/WorkedExampleBlock"
                },
                "propertyName": "type"
              },
              "oneOf": [
                {
                  "$ref": "#/components/schemas/TextBlock"
                },
                {
                  "$ref": "#/components/schemas/ListBlock"
                },
                {
                  "$ref": "#/components/schemas/CodeExampleBlock"
                },
                {
                  "$ref": "#/components/schemas/TableBlock"
                },
                {
                  "$ref": "#/components/schemas/WorkedExampleBlock"
                },
                {
                  "$ref": "#/components/schemas/CalloutBlock"
                },
                {
                  "$ref": "#/components/schemas/ImageBlock"
                },
                {
                  "$ref": "#/components/schemas/DiagramBlock"
                },
                {
                  "$ref": "#/components/schemas/AttachmentBlock"
                },
                {
                  "$ref": "#/components/schemas/RichTextBlock"
                },
                {
                  "$ref": "#/components/schemas/CodeVariantsBlock"
                }
              ]
            },
            "title": "Explanation",
            "type": "array"
          },
          "saved": {
            "title": "Saved",
            "type": "boolean"
          },
          "solution_revision": {
            "title": "Solution Revision",
            "type": "integer"
          }
        },
        "required": [
          "correct",
          "saved",
          "explanation",
          "solution_revision"
        ],
        "title": "CheckAndSaveResponse",
        "type": "object"
      },
...
      "CodeExampleBlock": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CodeExampleBlockData"
          },
          "type": {
            "const": "code_example",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "data"
        ],
        "title": "CodeExampleBlock",
        "type": "object"
      },
      "CodeExampleBlockData": {
        "additionalProperties": false,
        "properties": {
          "caption": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Caption"
          },
          "code": {
            "minLength": 1,
            "title": "Code",
            "type": "string"
          },
          "language": {
            "enum": [
              "python",
              "text"
            ],
            "title": "Language",
            "type": "string"
          }
        },
        "required": [
          "language",
          "code"
        ],
        "title": "CodeExampleBlockData",
        "type": "object"
      },
      "CodeVariant": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "maxLength": 10000,
            "minLength": 1,
            "title": "Code",
            "type": "string"
          },
          "label": {
            "maxLength": 40,
            "minLength": 1,
            "title": "Label",
            "type": "string"
          },
          "language": {
            "enum": [
              "python",
              "text"
            ],
            "title": "Language",
            "type": "string"
          }
        },
        "required": [
          "label",
          "language",
          "code"
        ],
        "title": "CodeVariant",
        "type": "object"
      },
      "CodeVariantsBlock": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CodeVariantsData"
          },
          "type": {
            "const": "code_variants",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "data"
        ],
        "title": "CodeVariantsBlock",
        "type": "object"
      },
      "CodeVariantsData": {
        "additionalProperties": false,
        "properties": {
          "variants": {
            "items": {
              "$ref": "#/components/schemas/CodeVariant"
            },
            "maxItems": 8,
            "minItems": 2,
            "title": "Variants",
            "type": "array"
          }
        },
        "required": [
          "variants"
        ],
        "title": "CodeVariantsData",
        "type": "object"
      },
...
      "DiagramBlock": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/DiagramData"
          },
          "type": {
            "const": "diagram",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "data"
        ],
        "title": "DiagramBlock",
        "type": "object"
      },
      "DiagramData": {
        "additionalProperties": false,
        "properties": {
          "accessible_description": {
            "maxLength": 10000,
            "minLength": 1,
            "title": "Accessible Description",
            "type": "string"
          },
          "alt": {
            "maxLength": 10000,
            "minLength": 1,
            "title": "Alt",
            "type": "string"
          },
          "caption": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Caption"
          },
          "height": {
            "exclusiveMinimum": 0.0,
            "maximum": 50000.0,
            "title": "Height",
            "type": "integer"
          },
          "pointers": {
            "items": {
              "$ref": "#/components/schemas/DiagramPointer"
            },
            "minItems": 1,
            "title": "Pointers",
            "type": "array"
          },
          "purpose": {
            "maxLength": 10000,
            "minLength": 1,
            "title": "Purpose",
            "type": "string"
          },
          "usage_id": {
            "pattern": "^[a-z0-9][a-z0-9_-]{0,119}$",
            "title": "Usage Id",
            "type": "string"
          },
          "width": {
            "exclusiveMinimum": 0.0,
            "maximum": 50000.0,
            "title": "Width",
            "type": "integer"
          }
        },
        "required": [
          "usage_id",
          "alt",
          "width",
          "height",
          "purpose",
          "accessible_description",
          "pointers"
        ],
        "title": "DiagramData",
        "type": "object"
      },
      "DiagramPointer": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "minLength": 1,
            "title": "Description",
            "type": "string"
          },
          "label": {
            "minLength": 1,
            "title": "Label",
            "type": "string"
          }
        },
        "required": [
          "label",
          "description"
        ],
        "title": "DiagramPointer",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ImageBlock": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/MediaData"
          },
          "type": {
            "const": "image",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "data"
        ],
        "title": "ImageBlock",
        "type": "object"
      },
      "InlineSpan": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "enum": [
              "text",
              "code",
              "formula"
            ],
            "title": "Kind",
            "type": "string"
          },
          "text": {
            "maxLength": 10000,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "text"
        ],
        "title": "InlineSpan",
        "type": "object"
      },
...
      "ListBlock": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ListBlockData"
          },
          "type": {
            "const": "list",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "data"
        ],
        "title": "ListBlock",
        "type": "object"
      },
      "ListBlockData": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Items",
            "type": "array"
          },
          "style": {
            "enum": [
              "ordered",
              "unordered"
            ],
            "title": "Style",
            "type": "string"
          }
        },
        "required": [
          "style",
          "items"
        ],
        "title": "ListBlockData",
        "type": "object"
      },
...
      "MediaData": {
        "additionalProperties": false,
        "properties": {
          "alt": {
            "maxLength": 10000,
            "minLength": 1,
            "title": "Alt",
            "type": "string"
          },
          "caption": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Caption"
          },
          "height": {
            "exclusiveMinimum": 0.0,
            "maximum": 50000.0,
            "title": "Height",
            "type": "integer"
          },
          "usage_id": {
            "pattern": "^[a-z0-9][a-z0-9_-]{0,119}$",
            "title": "Usage Id",
            "type": "string"
          },
          "width": {
            "exclusiveMinimum": 0.0,
            "maximum": 50000.0,
            "title": "Width",
            "type": "integer"
          }
        },
        "required": [
          "usage_id",
          "alt",
          "width",
          "height"
        ],
        "title": "MediaData",
        "type": "object"
      },
...
      "RichTextBlock": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/RichTextData"
          },
          "type": {
            "const": "rich_text",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "data"
        ],
        "title": "RichTextBlock",
        "type": "object"
      },
      "RichTextData": {
        "additionalProperties": false,
        "properties": {
          "spans": {
            "items": {
              "$ref": "#/components/schemas/InlineSpan"
            },
            "maxItems": 200,
            "minItems": 1,
            "title": "Spans",
            "type": "array"
          }
        },
        "required": [
          "spans"
        ],
        "title": "RichTextData",
        "type": "object"
      },
...
      "TableBlock": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/TableBlockData"
          },
          "type": {
            "const": "table",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "data"
        ],
        "title": "TableBlock",
        "type": "object"
      },
      "TableBlockData": {
        "additionalProperties": false,
        "properties": {
          "caption": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Caption"
          },
          "headers": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Headers",
            "type": "array"
          },
          "rows": {
            "items": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "minItems": 1,
            "title": "Rows",
            "type": "array"
          }
        },
        "required": [
          "headers",
          "rows"
        ],
        "title": "TableBlockData",
        "type": "object"
      },
...
      "TextBlock": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/TextBlockData"
          },
          "type": {
            "const": "text",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "data"
        ],
        "title": "TextBlock",
        "type": "object"
      },
      "TextBlockData": {
        "additionalProperties": false,
        "properties": {
          "markdown": {
            "minLength": 1,
            "title": "Markdown",
            "type": "string"
          }
        },
        "required": [
          "markdown"
        ],
        "title": "TextBlockData",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WorkedExampleBlock": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/WorkedExampleBlockData"
          },
          "type": {
            "enum": [
              "worked_example",
              "completion_exercise",
              "productive_failure_prompt"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "data"
        ],
        "title": "WorkedExampleBlock",
        "type": "object"
      },
      "WorkedExampleBlockData": {
        "additionalProperties": false,
        "properties": {
          "prompt": {
            "minLength": 1,
            "title": "Prompt",
            "type": "string"
          },
          "steps": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Steps",
            "type": "array"
          }
        },
        "required": [
          "prompt",
          "steps"
        ],
        "title": "WorkedExampleBlockData",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "infraege API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/progress/{context_kind}/{context_id}": {
      "delete": {
        "operationId": "delete_lesson_progress_api_progress__context_kind___context_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "context_kind",
            "required": true,
            "schema": {
              "title": "Context Kind",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "context_id",
            "required": true,
            "schema": {
              "title": "Context Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Lesson Progress",
        "tags": [
          "progress"
        ]
      }
    },
...
    "/api/tasks/{task_id}/check-and-save": {
      "post": {
        "operationId": "submit_answer_api_tasks__task_id__check_and_save_post",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckAndSaveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckAndSaveResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Answer",
        "tags": [
          "progress"
        ]
      }
    },
...
  }
}